use an handler for erros
