+++ pkg/autoscaling/execPolicy.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/autoscaling"
 )
 
@@ -16,26 +15,9 @@ func (client *Client) TriggerPolicy(asg *Autoscaling) error {
 		PolicyName:           aws.String(asg.PolicyName),
 	}
 
-	_, err := client.ExecutePolicy(input)
+	result, err := client.ExecutePolicy(input)
 	awsErr := checkAutoscalingError(err)
 
+	fmt.Println(result)
 	return awsErr
-}
-
-func checkAutoscalingError(err error) error {
-	if err != nil {
-		if aerr, ok := err.(awserr.Error); ok {
-			switch aerr.Code() {
-			case autoscaling.ErrCodeScalingActivityInProgressFault:
-				return fmt.Errorf("%v: %v", autoscaling.ErrCodeScalingActivityInProgressFault, aerr.Error())
-			case autoscaling.ErrCodeResourceContentionFault:
-				return fmt.Errorf("%v: %v", autoscaling.ErrCodeResourceContentionFault, aerr.Error())
-			default:
-				return fmt.Errorf("%v", aerr.Error())
-			}
-		} else {
-			return err
-		}
-	}
-	return nil
 }
